Fix VC cache setup/cleanup and test fails
[ERROR] Failures:
[ERROR]   TestVirtualCorpusCache.testCleanUpVC:152 expected:<3> but
was:<5>
[ERROR]   TestVirtualCorpusCache.testCollectionWithVCRefAndPubDate:476
expected:<2> but was:<1>
[ERROR]
TestVirtualCorpusCache.testNamedVCsAfterCorpusWithMissingDocs:407
[ERROR]
TestVirtualCorpusCache.testNamedVCsAfterQueryWithMissingDocs:335
[ERROR]   TestVirtualCorpusCache.testNestedNamedVCs:248
[ERROR]   TestVirtualCorpusCache.testReferToUncachedVC:88
[ERROR]   TestVirtualCorpusCache.testStoreUncachedVC:72
[ERROR]   TestVirtualCorpusCache.testUpdateCachedVC:112 expected:<2> but
was:<3>
[ERROR] Errors:
[ERROR]   TestVirtualCorpusCache.testCache:215 » IndexOutOfBounds Index
1 out of bounds for length 1
[INFO]
[ERROR] Tests run: 723, Failures: 8, Errors: 1, Skipped: 0

+++ src/main/java/de/ids_mannheim/korap/cache/VirtualCorpusCache.java
@@ -245,13 +245,22 @@ public class VirtualCorpusCache {
         map.clear();
 
         File vcCache = new File(CACHE_LOCATION + "/");
-        for (File vc : vcCache.listFiles()) {
-            for (File f : vc.listFiles()) {
-                if (f.exists()) {
-                    f.delete();
+        if (!vcCache.exists()) {
+            return;
+        }
+        File[] vcs = vcCache.listFiles();
+        if (vcs != null) {
+            for (File vc : vcs) {
+                File[] files = vc.listFiles();
+                if (files != null) {
+                    for (File f : files) {
+                        if (f.exists()) {
+                            f.delete();
+                        }
+                    }
                 }
+                vc.delete();
             }
-            vc.delete();
         }
         vcCache.delete();
     }
